test for init container restarts
Refactor tests and add an additional one to ensure init containers
don't restart upon regular container restart.
